:art: fix 2nd review correction

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By:  ctokoyod < ctokoyod@student.42tokyo.jp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/18 17:12:14 by  ctokoyod         #+#    #+#             */
-/*   Updated: 2024/01/05 14:08:15 by  ctokoyod        ###   ########.fr       */
+/*   Updated: 2024/01/15 21:40:45 by  ctokoyod        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*str != '\0')
